requirements.txt
streamlit
requests
jsonschema
ultralytics
pillow
torch
torchvision
opencv-python-headless  
huggingface_hub
